apps/backend/pyproject.toml
[project]
name = "helical-workbench-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apache-airflow-client>=3.1.6",
    "fastapi[standard]>=0.129.2",
    "pydantic-settings>=2.0",
    "requests>=2.32.5",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27",
    "mypy>=1.13",
    "pytest>=8.0",
    "pytest-mock>=3.14",
    "ruff>=0.9",
    "types-requests>=2.32",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.mypy]
python_version = "3.10"
strict = true
files = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I"]
